% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fttm_generate_data.R
\name{fttm_generate_ph_data}
\alias{fttm_generate_ph_data}
\title{Generate Data for FTTM Option 1}
\usage{
fttm_generate_ph_data(n = 500, m = 101, mu_c = 5, seed = 1)

fttm_generate_ph_data(n = 500, m = 101, mu_c = 5, seed = 1)
}
\arguments{
\item{n}{number of participants/sample size}

\item{m}{Number of functional time points}

\item{mu_c}{censoring times mean}

\item{seed}{seed for sampling/data generation}
}
\value{
A list of observed and true time observations, including censoring
times
}
\description{
Generate Data for FTTM Option 1
}
\details{
PO/\code{fttm_generate_po_data} - proportional odds
PH/\code{fttm_generate_ph_data} - proportional hazards
}
