% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fttm_generate_data.R
\name{fttm_generate_data}
\alias{fttm_generate_data}
\title{Generate Data for FTTM Option 1}
\usage{
fttm_generate_data(n = 500, m = 101, mu_c = 20, seed = 123)
}
\arguments{
\item{n}{number of observation}

\item{m}{rank of !! SOMETHING}

\item{mu_c}{censoring times mean}

\item{seed}{seed for sampling/data generation}
}
\value{
A list of observed and true time observations, including censoring
times
}
\description{
Generate Data for FTTM Option 1
}
